Added input validation for Team Name
This is so no team will end up being added without a team.
Also checks for a '-' to somewhat ensure the convention 'organization-team' is being followed.

An option to 'archive' old team members would be nice... i.e. move all 'rpi-mh' players to 'rpi-mh-old'. This could be reversed in the database in a pinch, and could be expanded to only archive players that are not found in the newly added roster, so there aren't constantly 30 people being added new every time (and no 'old' players that are duplicated and/or current players).