Player Title resizing updates
Auto-resize based on NHL draft pick or not.
Auto-expanding box for >1 character positions
Adjusted heights of everything in upper bar so it isn't too high above center.
Asjusted name to be less likely to overlap headshots if players have long wavy flow.
Fixed width of year so it doesn't overlap for 'So', etc.